Refactor file utility functions and reorganize file structure
